[batch] Rename old manage client file to match new task selector name
